fix * and add strjoin, substr, strtrim

+++ ft_strnstr.c
@@ -6,26 +6,25 @@
 /*   By: ngennaro <ngennaro@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 17:34:21 by ngennaro          #+#    #+#             */
-/*   Updated: 2022/11/15 18:15:29 by ngennaro         ###   ########lyon.fr   */
+/*   Updated: 2022/11/16 16:35:38 by ngennaro         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-#include <stdlib.h>
 
 char	*ft_strnstr(const char *str, const char *to_find, size_t size)
 {
-	int	i;
-	int	c;
+	size_t	i;
+	size_t	c;
 
 	i = 0;
 	c = 0;
 	if (ft_strlen(to_find) == 0)
-        return ((char *)str);
-    if (size == 0 || ft_strlen(str) == 0)
-        return (NULL);
-    if (ft_strlen(to_find) > ft_strlen(str))
-        return (NULL);
+		return ((char *)str);
+	if (size == 0 || ft_strlen(str) == 0)
+		return (NULL);
+	if (ft_strlen(to_find) > ft_strlen(str))
+		return (NULL);
 	while (str[i] != '\0' && i < size)
 	{
 		if (to_find[c] == str[i])
